+++ src/main/java/com/chrisrm/idea/actions/JoyAction.kt
@@ -4,7 +4,7 @@ import com.chrisrm.idea.MTConfig
 import com.chrisrm.idea.MTThemes
 import com.chrisrm.idea.actions.themes.BaseThemeAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import io.acari.DDLC.actions.ChibiOrchestrator
+import io.acari.DDLC.chibi.ChibiOrchestrator
 
 /**
  * Forged in the flames of battle by alex.
